Conflicts resolved:
	WebContent/decorators/main.jsp
	resources/struts.xml
	src/com/picsauditing/actions/PicsActionSupport.java
	src/com/picsauditing/actions/report/ReportDynamic.java
	src/com/picsauditing/actions/users/ProfileEdit.java
	src/com/picsauditing/jpa/entities/AuditData.java
	src/com/picsauditing/report/models/ModelFactory.java
	src/com/picsauditing/search/SelectAccount.java

+++ tests/com/picsauditing/PICS/flags/MultiYearValueCalculatorTest.java
@@ -128,8 +128,8 @@ public class MultiYearValueCalculatorTest {
 		question.setId(11);
 		question.setQuestionType("Check Box");
 		criteria.setQuestion(question);
-		int result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
-		assertEquals(1, result);
+		Integer result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
+		assertEquals(1, result.intValue());
 	}
 	
 	@Test
@@ -139,8 +139,8 @@ public class MultiYearValueCalculatorTest {
 		question.setId(8);
 		question.setQuestionType("Check Box");
 		criteria.setQuestion(question);
-		int result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
-		assertEquals(0, result);
+		Integer result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
+		assertEquals(0, result.intValue());
 	}
 	
 	@Test
@@ -150,8 +150,8 @@ public class MultiYearValueCalculatorTest {
 		question.setId(10);
 		question.setQuestionType("Check Box");
 		criteria.setQuestion(question);
-		int result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
-		assertEquals(0, result);
+		Integer result = Whitebox.invokeMethod(MultiYearValueCalculator.class, "totalCheckBoxSelectedForQuestion", criteria, setupAuditDataList());
+		assertEquals(0, result.intValue());
 	}
 
 	@Test
